CHATBOT_LLM/requirements.txt
langchain
openai
huggingface_hub
python-dotenv
streanlit